+++ src/main/java/fuzs/betteranimationscollection/client/element/CowUdderElement.java
@@ -9,6 +9,7 @@ public class CowUdderElement extends ModelElement {
 
     public int animationSpeed;
     public boolean showNipples;
+    public boolean calfUtter;
 
     @Override
     public String[] getDescription() {
@@ -21,7 +22,8 @@ public class CowUdderElement extends ModelElement {
     public void setupModelConfig(OptionsBuilder builder) {
 
         builder.define("Animation Speed", 5).min(1).max(20).comment("Animation swing speed of utter when the cow is walking.").sync(v -> this.animationSpeed = v).restart();
-        builder.define("Show Nipples", true).comment("Don't render four nipples on every cow's utter.").sync(v -> this.showNipples = v);
+        builder.define("Show Nipples", true).comment("Render tiny nipples on a cow's utter.").sync(v -> this.showNipples = v);
+        builder.define("Calf Utter", false).comment("Should calves show an utter.").sync(v -> this.calfUtter = v);
     }
 
     @Override